contributors/ttwf_shenzhen/css-shapes/sherlock/submitted/shape-outside-invalid-inset-rectangle-003.html

<!doctype html>
<html>
<head>
    <meta charset="UTF-8">
    <meta name="flags" content="dom">
    <meta name="assert" content="shape-outside declaration is invalid if inset-rectangle shape function has an invalid rx param">
    <title>CSS Shapes Test: Inset Rectangles - invalid rx</title>
    <link rel="author" title="Sherlock Zhong" href="mailto:sherlock.z0917@gmail.com">
    <link rel="help" href="http://www.w3.org/TR/css-shapes/#supported-basic-shapes">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="stylesheet" href="/resources/testharness.css">
    <style>
#shape{
    width: 300px;
    height: 300px;
    shape-outside: inset-rectangle(-50px, -50, 50px, 50px, 50, 50%);
}
    </style>
</head>
<body>

    <div id="log"></div>

    <div id="shape"></div>

    <script>
    var obj = document.getElementById('shape');
    var val = getComputedStyle(obj).shapeOutside;
    test(function() {assert_equals(val, "auto", "declaration should not be valid")});
    </script>

</body>
</html>
